requirements.txt
pyserial
# tkinter is included with standard Python installs on Windows; no pip package required.
